maint: drop useless casts
While reviewing all casts for an upcoming style consistency patch, I
noticed several casts that are semantically unnecessary in C:

- converting pointers to and from void* does not need a cast except
  for function pointers or when also casting away const (different
  from C++ where casts to or from void* are required for type safety);
  includes cases like python PyCapsule code which uses void*

- casting away const on a string literal (in C, string literals are
  typed 'char *', different from C++ where they are 'const char *';
  that said, C says it is still undefined behavior to modify their
  contents, so string literals should generally be only assigned to
  'const char *' variables or passed to functions that promise not to
  modify contents); includes cases like ocaml's struct
  custom_operations using 'const char *' as its member type (where
  casting away const just to re-add it is pointless)

- casts used to document what would otherwise be implicit type
  conversions when assigning to integers of a different type or
  passing function parameters not through varargs. While this can be
  useful for documentation purposes, it can also mask bugs when the
  cast uses the wrong intermediate type only to be implicitly
  converted back to the destination type, and it is often easier to
  use optional compiler warning flags to catch instances of unintended
  truncation or sign extension by implicit conversions.  For one of
  these, I added an assert that sysconf(_SC_NPROCESSORS_ONLN) is not
  going to overflow a 32-bit value.

Minimizing semantically unneeded casts lets the remaining casts stand
out for why they are necessary: casting aways const, casts between
pointer aliases (such as struct sockaddr*, char* vs unsigned char*,
...), and scalars passed through varargs in printf-like functions.

Message-Id: <20230524150003.384693-1-eblake@redhat.com>

+++ fuse/nbdfuse.c
@@ -415,7 +415,7 @@ main (int argc, char *argv[])
       handles_append (&nbd, h); /* reserved above, so can't fail */
     }
   }
-  connections = (unsigned) nbd.len;
+  connections = nbd.len;
   if (verbose)
     fprintf (stderr, "nbdfuse: connections=%u\n", connections);
 
@@ -424,7 +424,7 @@ main (int argc, char *argv[])
     fprintf (stderr, "%s\n", nbd_get_error ());
     exit (EXIT_FAILURE);
   }
-  size = (uint64_t) ssize;
+  size = ssize;
 
   /* If the remote NBD server is readonly, then act as if the '-r'
    * flag was given on the nbdfuse command line.
